refactor(server): move policy persistence into policy_store
